+++ User_Queries.py
@@ -3,6 +3,7 @@ import oracledb
 from flask import request, abort, jsonify
 import bcrypt
 
+import Permissions
 from Session import validate_session
 
 
@@ -14,6 +15,9 @@ def insert_user():
 
     if not all(key in data for key in ('username', 'password', 'email')):
         abort(400, 'Les données incomplètes pour l\'insertion')
+
+    if Permissions.has_permission(data.get("session_id"), "user.insert") is False:
+        return jsonify({"error": "Access denied"})
 
     cursor = connection.cursor()
 
@@ -50,6 +54,9 @@ def get_all_users():
     validation_response = validate_session(session_id, ip_address, user_agent)
     if 'error' in validation_response:
         return validation_response
+
+    if Permissions.has_permission(session_id, "user.get_data") is False:
+        return jsonify({"error": "Access denied"})
 
     cursor = connection.cursor()
 
@@ -96,6 +103,9 @@ def get_user():
     if 'error' in validation_response:
         return validation_response
 
+    if Permissions.has_permission(session_id, "user.get_data") is False:
+        return jsonify({"error": "Access denied"})
+
     cursor = connection.cursor()
 
     try:
@@ -140,6 +150,9 @@ def delete_user():
 
     if 'user_id' not in data:
         return jsonify({"message": "ID de l'utilisateur inexistant."}), 400
+
+    if Permissions.has_permission(session_id, "user.delete") is False:
+        return jsonify({"error": "Access denied"})
 
     user_id = data['user_id']
     cursor = connection.cursor()
